(PC-25817)[PRO] refactor: Replace SignUpJourneyBreadcrumb into Stepper.
