--- pro/src/components/SignupJourneyBreadcrumb/index.ts
@@ -1,1 +1,0 @@
-export { default as SignupJourneyBreadcrumb } from './SignupJourneyBreadcrumb'
