+++ pro/src/components/SignupJourneyStepper/SignupJourneyStepper.module.scss
