+++ pro/src/components/SignupJourneyStepper/index.ts
@@ -1,0 +1,1 @@
+export { default as SignupJourneyStepper } from './SignupJourneyStepper'
