+++ pro/src/screens/SignupJourneyForm/Offerer/Offerer.tsx
@@ -4,8 +4,8 @@ import { useNavigate } from 'react-router-dom'
 
 import { BannerInvisibleSiren } from 'components/Banner'
 import FormLayout from 'components/FormLayout'
-import { SIGNUP_JOURNEY_STEP_IDS } from 'components/SignupJourneyBreadcrumb/constants'
 import { OnboardingFormNavigationAction } from 'components/SignupJourneyFormLayout/constants'
+import { SIGNUP_JOURNEY_STEP_IDS } from 'components/SignupJourneyStepper/constants'
 import { useSignupJourneyContext } from 'context/SignupJourneyContext'
 import { Events } from 'core/FirebaseEvents/constants'
 import { FORM_ERROR_MESSAGE } from 'core/shared'
