+++ pro/src/screens/SignupJourneyForm/Validation/Validation.tsx
@@ -3,8 +3,8 @@ import { useNavigate } from 'react-router-dom'
 
 import { api } from 'apiClient/api'
 import { SaveNewOnboardingDataQueryModel, Target } from 'apiClient/v1'
-import { SIGNUP_JOURNEY_STEP_IDS } from 'components/SignupJourneyBreadcrumb/constants'
 import { OnboardingFormNavigationAction } from 'components/SignupJourneyFormLayout/constants'
+import { SIGNUP_JOURNEY_STEP_IDS } from 'components/SignupJourneyStepper/constants'
 import {
   DEFAULT_ACTIVITY_VALUES,
   useSignupJourneyContext,
